numpy-conversion-working Modified: - scratch.ipynb - ppo-vs-heuristic.ipynb - transition_mixin.py - euclidean_L2_embed.py

+++ waymo_agent/osmnx/euclidean_L2_embed.py
@@ -89,7 +89,7 @@ def interpolate_position_on_edge(
         interpolated = src + perc_done * (tgt - src)
         ret_df[val] = interpolated
 
-    return ret_df
+    return ret_df.astype(np.float32)
 
 
 def f_edges_start_end_node(
